Move config validation to own file
