+++ homeassistant/components/config/automation.py
@@ -2,11 +2,8 @@
 from collections import OrderedDict
 import uuid
 
-from homeassistant.components.automation import (
-    DOMAIN,
-    PLATFORM_SCHEMA,
-    async_validate_config_item,
-)
+from homeassistant.components.automation import DOMAIN, PLATFORM_SCHEMA
+from homeassistant.components.automation.config import async_validate_config_item
 from homeassistant.const import CONF_ID, SERVICE_RELOAD
 import homeassistant.helpers.config_validation as cv
